+++ config/owl-admin-kit.php
@@ -4,7 +4,7 @@ $publishMap = require __DIR__.'/publish-map.php';
 
 return [
 
-    'version' => '0.1.0',
+    'version' => '0.1.1',
 
     'name' => 'custom-admin-kit',